requirements.txt
tensorflow
matplotlib
numpy
sqlalchemy
psycopg2-binary
python-dotenv
alembic
passlib[bcrypt]
python-jose
pydantic
python-fasthtml
apscheduler
pandas